Add files via upload
This upload is the model file, the saved model, and the web app that I used from mtobeiyf.  The webapp/app.py is set up to use the saved_model.h5 file.  If you want to create a new model, change the parameters within the Image_Classifier.py and re-run the program from the terminal.  The program will create a saved model as "saved_model.h5".  Then move the saved model to the same directory as the webapp/app.py file.